chore: refactors stake flow logic
