+++ apps/ledger-live-desktop/src/renderer/screens/account/index.jsx
@@ -6,7 +6,6 @@ import { connect, useDispatch, useSelector } from "react-redux";
 import { withTranslation } from "react-i18next";
 import type { TFunction } from "react-i18next";
 import { Redirect } from "react-router";
-import { useHistory, useLocation } from "react-router-dom";
 import type { AccountLike, Account } from "@ledgerhq/types-live";
 import { SyncOneAccountOnMount } from "@ledgerhq/live-common/bridge/react/index";
 import { findCompoundToken } from "@ledgerhq/live-common/currencies/index";
@@ -41,7 +40,7 @@ import TokensList from "./TokensList";
 import CompoundBodyHeader from "~/renderer/screens/lend/Account/AccountBodyHeader";
 import useCompoundAccountEnabled from "~/renderer/screens/lend/useCompoundAccountEnabled";
 import { getBannerProps, AccountBanner } from "./AccountBanner";
-import perFamilyManageActions from "~/renderer/generated/AccountHeaderManageActions";
+import { useStartStakeFlow } from "~/renderer/screens/account/hooks/useStartStakeFlow";
 
 const mapStateToProps = (
   state,
@@ -92,32 +91,13 @@ const AccountPage = ({
     ? perFamilyAccountSubHeader[mainAccount.currency.family]
     : null;
   const bgColor = useTheme("colors.palette.background.paper");
-  const { state } = useLocation();
-  const history = useHistory();
 
   const isCompoundEnabled = useCompoundAccountEnabled(account, parentAccount);
 
   const [banner, setBanner] = useState({});
-
-  const manage = perFamilyManageActions[mainAccount.currency.family];
-  let manageList = [];
-  if (manage) {
-    const familyManageActions = manage({ account, parentAccount });
-    manageList = familyManageActions && familyManageActions.length > 0 ? familyManageActions : [];
-  }
+  useStartStakeFlow(account, parentAccount);
 
   const dispatch = useDispatch();
-
-  useEffect(() => {
-    const stakeAction = manageList.find(item => item.key === "Stake");
-
-    if (stakeAction && state?.startStake) {
-      history.replace();
-      stakeAction.onClick();
-    }
-    // ignoring manageList from dependency array
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [state, manageList]);
 
   useEffect(() => {
     if (mainAccount) {
